removed (int) from WEXITSTATUS

+++ src/mandatory/exec.c
@@ -6,7 +6,7 @@
 /*   By: mitasci <mitasci@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/19 18:04:04 by emgul             #+#    #+#             */
-/*   Updated: 2024/09/19 13:13:58 by mitasci          ###   ########.fr       */
+/*   Updated: 2024/09/19 14:54:38 by mitasci          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ static void	wait_for_pids(t_shell *shell, pid_t *pid, int cmdlen)
 	{
 		waitpid(pid[i], &exit_status, 0);
 		if (WIFEXITED(exit_status))
-			shell->last_exit_status = (int)WEXITSTATUS(exit_status);
+			shell->last_exit_status = WEXITSTATUS(exit_status);
 		i++;
 	}
 }
